+++ src/model/dfcxx/lib/dfcxx/typebuilders/builder.cpp
@@ -11,8 +11,8 @@
 namespace dfcxx {
 
 DFTypeImpl *TypeBuilder::buildFixed(FixedType::SignMode mode,
-                                    uint8_t intBits,
-                                    uint8_t fracBits) {
+                                    uint16_t intBits,
+                                    uint16_t fracBits) {
   return new FixedType(mode, intBits, fracBits);
 }
 
@@ -20,22 +20,24 @@ DFTypeImpl *TypeBuilder::buildBool() {
   return buildFixed(FixedType::SignMode::UNSIGNED, 1, 0);
 }
 
-DFTypeImpl *TypeBuilder::buildFloat(uint8_t expBits, uint8_t fracBits) {
+DFTypeImpl *TypeBuilder::buildFloat(uint16_t expBits, uint16_t fracBits) {
   return new FloatType(expBits, fracBits);
 }
 
-DFTypeImpl *TypeBuilder::buildShiftedType(DFTypeImpl *type, int8_t shift) {
+DFTypeImpl *TypeBuilder::buildShiftedType(DFTypeImpl *type, uint16_t shift) {
   if (type->isFixed()) {
     FixedType *casted = (FixedType *) type;
     return buildFixed(
         casted->getSign(),
-        uint8_t(int16_t(casted->getIntBits()) + shift),
-        casted->getFracBits());
+        casted->getIntBits() + shift,
+        casted->getFracBits()
+    );
   } else {
     FloatType *casted = (FloatType *) type;
     return buildFloat(
         casted->getExpBits(),
-        uint8_t(uint16_t(casted->getFracBits()) + shift));
+        casted->getFracBits() + shift
+    );
   }
 }
